Improve performance of permifrost grant runs
When SnowflakeConnector is initialised it creates an sqlalchemy
engine[1] which is quite costly. The docs for sqlalchemy recommend
creating an engine once per connection URL and using it throughout
the application[2]:

> the Engine is most efficient when created just once at the module
> level of an application, not per-object or per-function call.

To improve the performance of `permifrost grant` runs we now reuse the
same engine across all SQL grant queries that run. We've observed a
significant speedup, most likely due to the fact that we're no longer
initialising a new database connection for every SQL query that runs.

[1]: https://gitlab.com/gitlab-data/permifrost/-/blob/02f1c54ade4540d263549e9fb96da177a0f94daf/src/permifrost/core/permissions/utils/snowflake_connector.py#L17-51
[2]: https://docs.sqlalchemy.org/en/13/core/connections.html