Remove log-level from cli and use verbosity to determine log-level
It feels awkward to have both a verbose setting and a log-level. In the
future we should consider moving toward only one way of writing outputs,
but for now we can continue to use click for general user-interface
messages and logging for more detailed messages that might help with
debugging and performance testing.